+++ services/blockchain-connector/shared/app/sdk_v5/client.js
@@ -16,10 +16,10 @@
 const { Logger, Exceptions: { TimeoutException } } = require('lisk-service-framework');
 const { createWSClient } = require('@liskhq/lisk-api-client');
 
-const config = require('../../config');
-const delay = require('../utils/delay');
-const waitForIt = require('../utils/waitForIt');
-const Signals = require('../utils/signals');
+const config = require('../../../config');
+const delay = require('../../utils/delay');
+const waitForIt = require('../../utils/waitForIt');
+const Signals = require('../../utils/signals');
 
 const logger = Logger();
