+++ services/blockchain-connector/shared/app/sdk_v5/helper/encoder.js
