+++ services/blockchain-connector/shared/utils/fsUtils.js
