+++ services/blockchain-connector/shared/utils/signals.js
